chore(misc): move init telemetry into init-v2 handler
Addresses PR review: command-object.ts should not pull in ab-testing,
native (isAiAgent), utils helpers etc. at module load time — those run
for every nx command. Move the start/error telemetry and AI-mode error
output into the init-v2 outer handler, guarded by the `!inner` flag
so the downloaded-latest nx re-invocation does not double-record.
command-object.ts is now minimal again (yargs wiring only).

v1 path left untouched; slated for deletion in a follow-up.
