+++ packages/nx/src/command-line/init/command-object.ts
@@ -1,22 +1,6 @@
 import { Argv, CommandModule } from 'yargs';
 import { handleImport } from '../../utils/handle-import';
 import { parseCSV } from '../yargs-utils/shared-options';
-import { isAiAgent } from '../../native';
-import {
-  writeAiOutput,
-  buildErrorResult,
-  writeErrorLog,
-  determineErrorCode,
-} from './utils/ai-output';
-import { recordStat } from '../../utils/ab-testing';
-import { nxVersion } from '../../utils/versions';
-import { isCI } from '../../utils/is-ci';
-import { detectPackageManager } from '../../utils/package-manager';
-import {
-  extractErrorName,
-  readErrorStderr,
-  toErrorString,
-} from './implementation/utils';
 
 export const yargsInitCommand: CommandModule = {
   command: 'init',
@@ -47,88 +31,14 @@ export const yargsInitCommand: CommandModule = {
       throw error;
     });
 
-    const aiAgent = isAiAgent();
-
-    recordStat({
-      command: 'init',
-      nxVersion,
-      useCloud: false,
-      meta: {
-        type: 'start',
-        nodeVersion: process.versions.node,
-        os: process.platform,
-        packageManager: detectPackageManager(),
-        aiAgent,
-        isCI: isCI(),
-      },
-    });
-
-    try {
-      const useV2 = await isInitV2();
-      if (useV2) {
-        // v2 records its own complete event with richer context
-        await require('./init-v2').initHandler(args);
-      } else {
-        await require('./init-v1').initHandler(args);
-        await recordStat({
-          command: 'init',
-          nxVersion,
-          useCloud: false,
-          meta: {
-            type: 'complete',
-            nodeVersion: process.versions.node,
-            os: process.platform,
-            packageManager: detectPackageManager(),
-            aiAgent,
-            isCI: isCI(),
-          },
-        });
-      }
-      process.exit(0);
-    } catch (error) {
-      const errorMessage = toErrorString(error);
-      const errorCode = determineErrorCode(error);
-      // Append stderr tail (present when child-process errors ran with
-      // `stdio: 'pipe'`) so telemetry gets the real cause.
-      const stderr = readErrorStderr(error).trim();
-      const telemetryMessage = (
-        stderr
-          ? `${errorMessage} | stderr: ${stderr.slice(-250)}`
-          : errorMessage
-      ).slice(0, 500);
-      // Structured code for bucketing. Prefer Node's `e.code` (set on
-      // syscall failures); fall back to E-codes/ERR_* extracted from full
-      // stderr (npm/pnpm/yarn); then `error.name`.
-      const errorName = extractErrorName(error, stderr);
-
-      await recordStat({
-        command: 'init',
-        nxVersion,
-        useCloud: false,
-        meta: {
-          type: 'error',
-          errorCode,
-          errorName,
-          errorMessage: telemetryMessage,
-          aiAgent,
-          isCI: isCI(),
-          nodeVersion: process.versions.node,
-          os: process.platform,
-          packageManager: detectPackageManager(),
-        },
-      });
-
-      // Output structured error for AI agents
-      if (aiAgent) {
-        const errorLogPath = writeErrorLog(error);
-        writeAiOutput(buildErrorResult(errorMessage, errorCode, errorLogPath));
-      } else {
-        // Ensure the cursor is always restored just in case the user has bailed during interactive prompts
-        // Skip for AI agents to avoid corrupting NDJSON output
-        process.stdout.write('\x1b[?25h');
-      }
-      process.exit(1);
+    const useV2 = await isInitV2();
+    if (useV2) {
+      await require('./init-v2').initHandler(args);
+    } else {
+      // v1 path retained for `NX_ADD_PLUGINS=false`; slated for removal.
+      await require('./init-v1').initHandler(args);
     }
+    process.exit(0);
   },
 };
