+++ packages/nx/src/command-line/init/init-v2.ts
@@ -23,10 +23,13 @@ import { addNxToAngularCliRepo } from './implementation/angular';
 import { generateDotNxSetup } from './implementation/dot-nx/add-nx-scripts';
 import {
   createNxJsonFile,
+  extractErrorName,
   initCloud,
   isMonorepo,
   printFinalMessage,
+  readErrorStderr,
   setNeverConnectToCloud,
+  toErrorString,
   updateGitIgnore,
 } from './implementation/utils';
 import { ensurePackageHasProvenance } from '../../utils/provenance';
@@ -64,32 +67,100 @@ export async function initHandler(
   options: InitArgs,
   inner = false
 ): Promise<void> {
-  // Use environment variable to force local execution
-  if (process.env.NX_USE_LOCAL === 'true' || inner) {
-    return await initHandlerImpl(options);
-  }
-
-  let cleanup: () => void | undefined;
-  try {
-    await ensurePackageHasProvenance('nx', 'latest');
-    const packageInstallResults = installPackageToTmp('nx', 'latest');
-    cleanup = packageInstallResults.cleanup;
-
-    let modulePath = require.resolve('nx/src/command-line/init/init-v2.js', {
-      paths: [packageInstallResults.tempDir],
+  // Only the outermost CLI invocation records start/error telemetry. When
+  // the downloaded-latest nx re-invokes us with `inner=true` its own outer
+  // wrapper already recorded these events.
+  const selfRecord = !inner;
+  const aiAgent = isAiAgent();
+  const baseMeta = {
+    nodeVersion: process.versions.node,
+    os: process.platform,
+    packageManager: detectPackageManager(),
+    aiAgent,
+    isCI: isCI(),
+  };
+  if (selfRecord) {
+    recordStat({
+      command: 'init',
+      nxVersion,
+      useCloud: false,
+      meta: { type: 'start', ...baseMeta },
     });
-
-    const module = await handleImport(modulePath);
-    const result = await module.initHandler(options, true);
-    cleanup();
-    return result;
-  } catch (error) {
-    if (cleanup) {
-      cleanup();
-    }
-    // Fall back to local implementation
-    return initHandlerImpl(options);
   }
+
+  try {
+    if (process.env.NX_USE_LOCAL === 'true' || inner) {
+      return await initHandlerImpl(options);
+    }
+
+    let cleanup: () => void | undefined;
+    try {
+      await ensurePackageHasProvenance('nx', 'latest');
+      const packageInstallResults = installPackageToTmp('nx', 'latest');
+      cleanup = packageInstallResults.cleanup;
+
+      let modulePath = require.resolve('nx/src/command-line/init/init-v2.js', {
+        paths: [packageInstallResults.tempDir],
+      });
+
+      const module = await handleImport(modulePath);
+      const result = await module.initHandler(options, true);
+      cleanup();
+      return result;
+    } catch {
+      if (cleanup) cleanup();
+      // Fall back to local implementation
+      return await initHandlerImpl(options);
+    }
+  } catch (error) {
+    if (selfRecord) {
+      await recordInitError(error, aiAgent, baseMeta);
+      // recordInitError terminates the process on the CLI path; the throw
+      // below is unreachable in practice but kept for type-level safety.
+    }
+    throw error;
+  }
+}
+
+async function recordInitError(
+  error: unknown,
+  aiAgent: boolean,
+  baseMeta: Record<string, string | boolean>
+): Promise<void> {
+  const errorMessage = toErrorString(error);
+  const errorCode = determineErrorCode(error);
+  // Append stderr tail (attached when child-process errors ran with
+  // `stdio: 'pipe'`) so telemetry carries the real cause.
+  const stderr = readErrorStderr(error).trim();
+  const telemetryMessage = (
+    stderr ? `${errorMessage} | stderr: ${stderr.slice(-250)}` : errorMessage
+  ).slice(0, 500);
+  // Structured code for bucketing. Prefer Node's `e.code`; fall back to
+  // E-code/ERR_* tokens extracted from stderr; then `error.name`.
+  const errorName = extractErrorName(error, stderr);
+
+  await recordStat({
+    command: 'init',
+    nxVersion,
+    useCloud: false,
+    meta: {
+      type: 'error',
+      errorCode,
+      errorName,
+      errorMessage: telemetryMessage,
+      ...baseMeta,
+    },
+  });
+
+  if (aiAgent) {
+    const errorLogPath = writeErrorLog(error);
+    writeAiOutput(buildErrorResult(errorMessage, errorCode, errorLogPath));
+  } else {
+    // Restore the cursor in case the user bailed during an interactive
+    // prompt. Skip for AI agents — it would corrupt NDJSON output.
+    process.stdout.write('\x1b[?25h');
+  }
+  process.exit(1);
 }
 
 async function initHandlerImpl(options: InitArgs): Promise<void> {
